1、文件上传大小限制更改为500MB 2、pc单个文件上传接口编写 3、完成移动文件，支持递归移动 4、移动文件限制，当前目录不可移动，目标路径是已选文件的子文件夹时不可移动 5、文件重命名功能 6、文件删除功能 7、前端点击上传文件，支持多个文件上传，带进度条。 8、前端拖拽多个文件到文件管理页面区域自动上传，带进度条。 9、单个文件同步到手机功能 10、多个文件批量同步到手机功能 11、根据配置同步路径，一键同步Web端文件到手机端(文件夹不会递归同步) 12、根据配置同步路径，一键同步手机端文件到Web端(文件夹不会递归同步) 13、自动同步web同步路径功能，开启后，点击面包屑导航，自动刷新web端同步路径

+++ src/main/java/com/zjh/zxw/service/impl/AttachmentInfoServiceImpl.java
@@ -259,9 +259,17 @@ public class AttachmentInfoServiceImpl implements AttachmentInfoService {
 
 
     @Override
-    public Boolean reNameFile(String oldFileName, String newFileName) {
-        File file = new File(this.getRootPath() + oldFileName);
-        File newFile = new File(this.getRootPath() + newFileName);
+    public Boolean reNameFile(String oldFilePathName, String newFilePathName) {
+        String prePath = this.getRootPath();
+        if(!oldFilePathName.contains(prePath) || !newFilePathName.contains(prePath)){
+            throw new BusinessException("非指定目录,不可进行操作");
+        }
+
+        File file = new File(oldFilePathName);
+        File newFile = new File(newFilePathName);
+        if(newFile.exists()){
+            throw new BusinessException("当前名称已存在");
+        }
         boolean renameTo = file.renameTo(newFile);
         return renameTo;
     }
@@ -279,6 +287,7 @@ public class AttachmentInfoServiceImpl implements AttachmentInfoService {
     @Override
     public Boolean createFolder(String folderName) {
         String folderPath = this.getRootPath() + folderName;
+        folderPath = this.getTargetFilePathNoExit(folderPath,true);
         FileUtil.createFolder(folderPath);
         return true;
     }
